infinity/smart-build: do it other way around (esthetics)
